% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate-boundaries.R
\name{locate_boundaries}
\alias{locate_boundaries}
\title{Locate boundaries around changes in a vector}
\usage{
locate_boundaries(x)
}
\arguments{
\item{x}{\verb{[integer / double]}

The vector to locate boundaries in. Can only be an integer or double
vector.}
}
\value{
A two column data frame with double columns \code{start} and \code{stop}.
}
\description{
\code{locate_boundaries()} returns the running start and stop positions
identifying sections of unchanged values in \code{x}.
}
\details{
\code{1} is always the first \code{start} position, and \code{length(x)} is always the last
\code{stop} position.
}
\examples{
locate_boundaries(c(2, 2, 2, 3, 5, 5))
}
