% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate-changes.R
\name{locate_changes}
\alias{locate_changes}
\title{Locate change points in a vector}
\usage{
locate_changes(x)
}
\arguments{
\item{x}{\verb{[integer / double]}

The vector to locate changes in. Can only be an integer or double vector.}
}
\value{
An integer vector mapping to the last value of \code{x} before a change is
detected.
}
\description{
\code{locate_changes()} determines where an integer or double vector changes
values. It returns an integer vector containing the positions of the last
element before the change. These are the "stop points" used in
\code{\link[=locate_boundaries]{locate_boundaries()}} and \code{\link[=warp_boundaries]{warp_boundaries()}}. The position of the last
element of \code{x} is always returned at the end of the vector.
}
\examples{
locate_changes(c(1, 1, 2))

locate_changes(c(1.5, 1, 2))

# NA and NaN values are treated as being equivalent with themselves
# but not with each other
locate_changes(c(1, NA_real_, NA_real_, NaN, NaN))
}
