% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timewarp-package.R
\docType{package}
\name{timewarp-package}
\alias{timewarp}
\alias{timewarp-package}
\title{timewarp: Group Dates and Datetimes}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/timewarp}
  \item Report bugs at \url{https://github.com/DavisVaughan/timewarp/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
