% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp-package.R
\docType{package}
\name{warp-package}
\alias{warp}
\alias{warp-package}
\title{warp: Group Dates and Datetimes}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/warp}
  \item Report bugs at \url{https://github.com/DavisVaughan/warp/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
