% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{warp_boundaries}
\alias{warp_boundaries}
\title{Locate period boundaries for a date vector}
\usage{
warp_boundaries(x, by = "year", every = 1L, origin = NULL)
}
\arguments{
\item{x}{\verb{[Date / POSIXct / POSIXlt]}

The vector to compute time boundaries for.}

\item{by}{\verb{[character(1)]}

A string defining the period to group by. Valid inputs are:
\itemize{
\item \code{"year"}
\item \code{"quarter"}
\item \code{"month"}
\item \code{"week"}
\item \code{"day"}
\item \code{"hour"}
\item \code{"minute"}
\item \code{"second"}
}}

\item{every}{\verb{[positive integer(1)]}

The number of \code{by} periods to lump together when constructing the groups.
For example, with the default \code{origin} and \code{by = "year"}, \code{every = 2}
would put the years \code{1970} and \code{1971} in the same group.}

\item{origin}{\verb{[Date(1) / POSIXct(1) / POSIXlt(1)]}

The reference value for the groups to be computed from. This is
particularly important when \code{every > 1} and you need to define when the
"first" event was to start counting from.}
}
\description{
\code{warp_boundaries()} detects a change in time period along \code{x}, for example,
rolling from one month to the next. It returns the start and stop positions
for each contiguous period chunk in \code{x}.
}
\examples{
x <- as.Date("1970-01-01") + -4:5
x

# Boundaries by month
warp_boundaries(x, "month")

# Bound by every 5 days, relative to "1970-01-01"
# Creates boundaries of:
# [1969-12-27, 1970-01-01)
# [1970-01-01, 1970-01-06)
# [1970-01-06, 1970-01-11)
warp_boundaries(x, "day", every = 5)

# Bound by every 5 days, relative to the smallest value in our vector
origin <- min(x)
origin

# Creates boundaries of:
# [1969-12-28, 1970-01-02)
# [1970-01-02, 1970-01-07)
warp_boundaries(x, "day", every = 5, origin = origin)
}
