% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{warp_boundary}
\alias{warp_boundary}
\title{Locate period boundaries for a date vector}
\usage{
warp_boundary(x, by = "year", every = 1L, origin = NULL)
}
\arguments{
\item{x}{\verb{[Date / POSIXct / POSIXlt]}

A date time vector.}

\item{by}{\verb{[character(1)]}

A string defining the period to group by. Valid inputs are:

\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, \code{"day"}, \code{"hour"}, \code{"minute"},
\code{"second"}, \code{"millisecond"}}

\item{every}{\verb{[positive integer(1)]}

The number of \code{by} periods to group together.

For example, if \code{by = "year"} and \code{every} is set to \code{2}, then the years
1970 and 1971 would be placed in the same group.}

\item{origin}{\verb{[Date(1) / POSIXct(1) / POSIXlt(1)]}

The reference date time value. The default when left as \code{NULL} is the
Unix epoch of \verb{1970-01-01 00:00:00 UTC}.

This is used for two purposes:
\itemize{
\item Defining the anchor to count from when \code{every > 1}.
\item Aligning the time zone with the input. When the input vector and the
\code{origin} have altering time zones, a warning is issued and the input
is coerced to the time zone of the \code{origin}. It is highly advised to
provide your own \code{origin} if your input is a POSIXct.
}}
}
\value{
A two column data frame with the columns \code{start} and \code{stop}. Both are
double vectors representing boundaries of the date time groups.
}
\description{
\code{warp_boundary()} detects a change in time period along \code{x}, for example,
rolling from one month to the next. It returns the start and stop positions
for each contiguous period chunk in \code{x}.
}
\details{
The stop positions are just the \code{\link[=warp_change]{warp_change()}} values, and the start
positions are computed from these.
}
\examples{
x <- as.Date("1970-01-01") + -4:5
x

# Boundaries by month
warp_boundary(x, "month")

# Bound by every 5 days, relative to "1970-01-01"
# Creates boundaries of:
# [1969-12-27, 1970-01-01)
# [1970-01-01, 1970-01-06)
# [1970-01-06, 1970-01-11)
warp_boundary(x, "day", every = 5)

# Bound by every 5 days, relative to the smallest value in our vector
origin <- min(x)
origin

# Creates boundaries of:
# [1969-12-28, 1970-01-02)
# [1970-01-02, 1970-01-07)
warp_boundary(x, "day", every = 5, origin = origin)
}
