% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeslide-package.R
\docType{package}
\name{timeslide-package}
\alias{timeslide}
\alias{timeslide-package}
\title{timeslide: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/timeslide}
  \item Report bugs at \url{https://github.com/DavisVaughan/timeslide/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
