% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp-package.R
\docType{package}
\name{warp-package}
\alias{warp}
\alias{warp-package}
\title{warp: Group Dates}
\description{
Tooling to group dates by a variety of periods including: yearly, monthly, by second, by week of the month, and more. The groups are defined in such a way that they also represent the distance between dates in terms of the period. This extracts valuable information that can be used in further calculations that rely on a specific temporal spacing between observations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/warp}
  \item \url{https://davisvaughan.github.io/warp/}
  \item Report bugs at \url{https://github.com/DavisVaughan/warp/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
