% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{warp_change}
\alias{warp_change}
\title{Detect changes in a date time vector}
\usage{
warp_change(x, period = "year", every = 1L, origin = NULL)
}
\arguments{
\item{x}{\verb{[Date / POSIXct / POSIXlt]}

A date time vector.}

\item{period}{\verb{[character(1)]}

A string defining the period to group by. Valid inputs are:

\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, \code{"day"}, \code{"hour"}, \code{"minute"},
\code{"second"}, \code{"millisecond"}}

\item{every}{\verb{[positive integer(1)]}

The number of \code{period}s to group together.

For example, if \code{period = "year"} and \code{every} is set to \code{2}, then the years
1970 and 1971 would be placed in the same group.}

\item{origin}{\verb{[Date(1) / POSIXct(1) / POSIXlt(1)]}

The reference date time value. The default when left as \code{NULL} is the
Unix epoch of \verb{1970-01-01 00:00:00 UTC}.

This is used for two purposes:
\itemize{
\item Defining the anchor to count from when \code{every > 1}.
\item Aligning the time zone with the input. When the input vector and the
\code{origin} have altering time zones, a warning is issued and the input
is coerced to the time zone of the \code{origin}. It is highly advised to
provide your own \code{origin} if your input is a POSIXct.
}}
}
\value{
A double vector of locations right before a change.
}
\description{
\code{warp_change()} detects changes at the \code{period} level. It returns the
locations of the value just before a change. The location of the last
value in \code{x} is always returned at the end.
}
\examples{
x <- as.Date("2019-01-01") + 0:5

warp_change(x, period = "day", every = 2)

warp_change(x, period = "day", every = 2, origin = as.Date("2019-01-01"))
}
